pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "forecastbench-rationale"
version = "0.1.0"
description = "Rationale scoring tools for ForecastBench data."
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
analyze-questions = "forecastbench_rationale.analyze_questions:main"
select-questions = "forecastbench_rationale.select_questions:main"
score-rationales = "forecastbench_rationale.score_rationales:main"
validate-scores = "forecastbench_rationale.validate_scores:main"

[tool.setuptools.packages.find]
where = ["src"]
